dummy.js
public abstract java.util.List<java.lang.String> getRaw()

Retrieves a read-only list of the raw arguments. This list may be empty, but is never null. In the case of a standalone application, it is the ordered list of arguments specified on the command line. In the case of an applet or WebStart application, it includes unnamed parameters as well as named parameters. For named parameters, each <name,value> pair is represented as a single argument of the form: "--name=value".

Returns:
    a read-only list of raw application arguments