platform.js
Make it official
Community Documentation

    Community Docs Home
    Develop for Drupal
    Theming Guide
    Glossary
    Contribute to Docs

Streaming content from a Flash Media Server with SWF Tools

Last updated October 27, 2010. Created on March 11, 2009.
Edited by elvis2, Stuart Greenfield. Log in to edit this page.

As of SWF Tools 6.x-2.x there is early support for streaming media from Flash Media Servers (such as Adobe's Flash Media Server, Red5 open source server, or Wowza).

This page is only partial at the moment, but it is intended to provide something of a reference guide if you want to experiment with the features. Note that setting up a media server can be tricky, and this page can't really help you out with that part - it is intended to help you use SWF Tools with a working media server!

At present, streaming media is support in the Wijering 4 and FlowPlayer 3 modules. Both can handle single media streams, whilst FlowPlayer3 can also handle streamed playlists.

The examples below show how to stream media using the swf filter syntax, and an example for PHP is also provided as it can be a little bit confusing as to where to pass the various parameters.
[swf] input filter

In summary, all you need to do is add the parameter stream="rtmp://server/application" to your regular [swf] string. Because for the media servers you don't always include the extension, or you might be streaming a format other than flv, you should also tell SWF Tools to treat the file as a video stream by setting action="swftools_flv_display" or action="swftools_flv_display_list".

The best way to understand it is with some examples...

To stream a single flv video file:
[swf file="sample" stream="rtmp://server/app" action="swftools_flv_display"]

Note that the extension is not included on the filename for an flv stream.

To stream a single f4v:
[swf file="mp4:myVideo.f4v" stream="rtmp://server/app" action="swftools_flv_display"]

Note in this case that the file type is specified by the prefix mp4:, and the extension is included.

To stream a media playlist (FlowPlayer3 only):
[swf files="videoOne&&videoTwo" stream="rtmp://server/app" action="swftools_flv_display_list" playlist="true"]

You can even pick up a live video stream. For example, using the included live service with Flash Media Server you can use