Add files via upload
return myapp;You can create the user interface elements in any order. The order in which you add elements to the objects determines the display order. For example, if you create a VerticalPanel object, and add elements a, b and c to the panel, the elements are displayed from top to bottom in that order. However, you can create a, b, and c in any order. Creating the elements and adding them to your application are separate steps requiring separate instructions.